requirements.txt
python-telegram-bot==22.6
google-genai>=0.5.0
apscheduler==3.10.4
aiosqlite==0.19.0
python-dotenv==1.0.0
pytz==2024.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0